pyproject.toml
[project]
name = "enterprise-rag-api"
version = "0.1.0"
description = "Enterprise-grade RAG backend"
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "psycopg[binary]>=3.2.0",
  "pgvector>=0.3.5",
  "python-jose[cryptography]>=3.3.0",
  "pwdlib[argon2]>=0.2.0",
  "email-validator>=2.2.0",
  "llama-index>=0.11.0",
  "sentence-transformers>=3.0.0",
  "ragas>=0.2.0",
  "pypdf>=5.0.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "orjson>=3.10.0",
  "structlog>=24.4.0",
  "tenacity>=9.0.0",
  "pandas>=2.2.0",
  "matplotlib>=3.9.0",
  "pytest>=8.0.0",
  "python-multipart>=0.0.9",
  "datasets>=2.20.0",
  "langchain-openai>=0.2.0",
  "pytest>=8.0.0",
  "pytest-mock>=3.14.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"